pyproject.toml
[project]
name = "session6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.56.0",
    "mcp>=1.25.0",
    "pillow>=12.0.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
